requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
django-rest-auth==0.9.5
django-seed==0.2.2
djangorestframework==3.12.2
Faker==6.6.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
